Analytics v1 done

+++ Platibus.Web/DataServices/Models/User/UserNameAndSalary.cs
@@ -3,9 +3,9 @@ namespace Platibus.Web.DataServices.Models.User
     public class UserNameAndSalary
     {
         public string name { get; set; }
-        public int salary { get; set; }
+        public double salary { get; set; }
 
-        public UserNameAndSalary(string _name, int _salary)
+        public UserNameAndSalary(string _name, double _salary)
         {
             name = _name;
             salary = _salary;
